+++ 10up-experience.php
@@ -2,7 +2,7 @@
 /**
  * Plugin Name: 10up Experience
  * Description: The 10up Experience plugin configures WordPress to better protect and inform clients, aligned to 10up’s best practices.
- * Version:     1.7
+ * Version:     1.7.2
  * Author:      10up
  * Author URI:  https://10up.com
  * License:     GPLv2 or later
@@ -16,16 +16,29 @@ namespace TenUpExperience;
 
 use Puc_v4_Factory;
 
-define( 'TENUP_EXPERIENCE_VERSION', '1.6.2' );
+define( 'TENUP_EXPERIENCE_VERSION', '1.7.2' );
 define( 'TENUP_EXPERIENCE_DIR', __DIR__ );
 define( 'TENUP_EXPERIENCE_FILE', __FILE__ );
 
-if ( ! file_exists( __DIR__ . '/vendor/autoload.php' ) ) {
-	wp_die( esc_html__( 'This plugin requires vendor/autoload.php which is missing. The file should be bundled with the plugin and is committed to the repository.' ) );
-}
+require_once __DIR__ . '/vendor/yahnis-elsts/plugin-update-checker/plugin-update-checker.php';
 
-require_once __DIR__ . '/vendor/autoload.php';
 require_once __DIR__ . '/includes/utils.php';
+
+spl_autoload_register( function( $class_name ) {
+	$path_parts = explode( '\\', $class_name );
+
+	if ( ! empty( $path_parts ) ) {
+		$package = $path_parts[0];
+
+		unset( $path_parts[0] );
+
+		if ( 'TenUpExperience' === $package ) {
+			require_once __DIR__ . '/includes/classes/' . implode( '/', $path_parts ) . '.php';
+		} elseif ( 'ZxcvbnPhp' === $package ) {
+			require_once __DIR__ . '/vendor/bjeavons/zxcvbn-php/src/' . implode( '/', $path_parts ) . '.php';
+		}
+	}
+} );
 
 $tenup_plugin_updater = Puc_v4_Factory::buildUpdateChecker(
 	'https://github.com/10up/10up-experience/',
@@ -52,12 +65,31 @@ $network_activated = Utils\is_network_activated( plugin_basename( __FILE__ ) );
 
 define( 'TENUP_EXPERIENCE_IS_NETWORK', (bool) $network_activated );
 
-AdminCustomizations\Customizations::instance()->setup();
+if ( ! defined( 'TENUP_DISABLE_BRANDING' ) || ! TENUP_DISABLE_BRANDING ) {
+	AdminCustomizations\Customizations::instance()->setup();
+}
+
 API\API::instance()->setup();
-Authentication\Passwords::instance()->setup();
+Authentication\Usernames::instance()->setup();
 Authors\Authors::instance()->setup();
 Gutenberg\Gutenberg::instance()->setup();
+Headers\Headers::instance()->setup();
 Plugins\Plugins::instance()->setup();
 PostPasswords\PostPasswords::instance()->setup();
 SupportMonitor\Monitor::instance()->setup();
 SupportMonitor\Debug::instance()->setup();
+Notifications\Welcome::instance()->setup();
+
+/**
+ * We load this later to make sure there are no conflicts with other plugins.
+ */
+add_action( 'plugins_loaded', function() {
+	Authentication\Passwords::instance()->setup();
+} );
+
+/**
+ * Disable plugin/theme editor
+ */
+if ( ! defined( 'DISALLOW_FILE_EDIT' ) ) {
+	define( 'DISALLOW_FILE_EDIT', true );
+}